Find me button added to fetch users location
