+++ src/App.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import Header from './components/header/Header.jsx';
 import Homepage from './components/homepage/Homepage';
